React file structure. Builds in heroku.
