+++ .gitignore
@@ -1,4 +1,5 @@
 cv.txt
+build/
 
 #
 # Following copied from a template .gitignore
