+++ Procfile
@@ -1,1 +1,1 @@
-web: npm start
+web: node --inspect=5858 server
