+++ react/public/manifest.json
