+++ react/src/index.js
@@ -1,0 +1,8 @@
+import React from 'react'
+import ReactDOM from 'react-dom'
+import './index.css'
+import App from './components/App'
+// import registerServiceWorker from './registerServiceWorker' //Caching is disabled for now.
+
+ReactDOM.render(<App />, document.getElementById('root'))
+// registerServiceWorker()
